requirements.txt
# Discord
discord.py>=2.3.2

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Japanese Text Processing
fugashi>=1.3.0
unidic-lite>=1.0.8
jaconv>=0.3.4

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1

# Optional: for future enhancements
# pykakasi>=2.2.1  # furigana generation